breakout manual benchmarks into seperate file
